Redundancy

+++ payCalculator.py
@@ -5,7 +5,7 @@ def calculatePay():
     # This first line is provided for you
     hrs = float(input("Enter Hours:"))
     rate = float(input("Enter Rate:"))
-    pay = hrs * rate
+    pay = float(hrs * rate)
     print(pay)
     # end assignment
